+++ esmecata/core/eggnog.py
@@ -431,6 +431,7 @@ def annotate_with_eggnog(input_folder, output_folder, eggnog_database_path, nb_c
     options['tool_dependencies']['python_package'] = {}
     options['tool_dependencies']['python_package']['Python_version'] = sys.version
     options['tool_dependencies']['python_package']['esmecata'] = esmecata_version
+    options['tool_dependencies']['python_package']['pandas'] = pd.__version__
     options['tool_dependencies']['eggnog_mapper'] = get_eggnog_version()
 
     esmecata_metadata = {}