+++ esmecata/gseapy/gseapy_orsum.py
@@ -16,6 +16,9 @@ import csv
 import os
 import sys
 import pandas as pd
+import datetime
+import time
+import json
 import pronto
 import logging
 import subprocess
@@ -24,7 +27,10 @@ import urllib.request
 import gseapy
 
 from Bio.ExPASy import Enzyme
+from matplotlib import __version__ as matplotlib_version
 from esmecata.core.proteomes import get_taxon_obs_name
+from esmecata import __version__ as esmecata_version
+from Bio import __version__ as biopython_version
 
 logger = logging.getLogger(__name__)
 
@@ -108,6 +114,26 @@ def run_orsum(annotation_file_gmt_file, enrichr_module_phylum_output, orsum_outp
     subprocess.call(['orsum.py', '--gmt', annotation_file_gmt_file, '--files', *input_files, '--outputFolder', orsum_output_folder])
 
 
+def get_orsum_version():
+    """Returns version of orsum.
+
+    Returns:
+        orsum_version (str): version of orsum.
+    """
+    orsum_version = None
+    response = subprocess.Popen(['orsum.py', '-v'], stdout=subprocess.PIPE, start_new_session=True, universal_newlines="")
+    for orsum_line in response.stdout:
+        orsum_line_decoded = orsum_line.decode('utf-8')
+        orsum_version = orsum_line_decoded.strip('\n')
+
+    if orsum_version is None:
+        logger.critical('|EsMeCaTa|gseapy_taxon| esmecata could not find the version of orsum.')
+        logger.critical('|EsMeCaTa|gseapy_taxon| It is possibly an issue with the installation of orsum (maybe it is not in the PATH). Or it can be due to a change in the output of orsum.py -v command.')
+        sys.exit()
+
+    return orsum_version
+
+
 def taxon_rank_annotation_enrichment(annotation_folder, output_folder, enzyme_data_file=None, go_basic_obo_file=None, taxon_rank='phylum'):
     """ Run an enrichment analysis on taxon from annotation results of esmecata using gseapy.
     Then filter this list with orsum.
@@ -119,12 +145,34 @@ def taxon_rank_annotation_enrichment(annotation_folder, output_folder, enzyme_da
         go_basic_obo_file (str): path to Gene Ontology go-basic.obo file, if not given, download it
         taxon_rank (str): taxon rank to cluster the observation name together (by default, phylum)
     """
+    starttime = time.time()
+    logger.info('|EsMeCaTa|gseapy_taxon| Begin enrichment analysis.')
     taxon_ranks = ['species', 'genus', 'family', 'order', 'class', 'phylum', 'kingdom', 'superkingdom']
 
     if taxon_rank != 'phylum':
         if taxon_rank not in taxon_ranks:
             logger.critical('Incorrect taxon given {0}, possible ranks are: {1}'.format(taxon_rank, ','.join(taxon_ranks)))
             sys.exit()
+
+    # Get metadata associated with run.
+    options = {'annotation_folder': annotation_folder, 'output_folder': output_folder, 'enzyme_data_file': enzyme_data_file,
+                'go_basic_obo_file': go_basic_obo_file, 'taxon_rank': taxon_rank}
+
+    options['tool_dependencies'] = {}
+    options['tool_dependencies']['python_package'] = {}
+    options['tool_dependencies']['python_package']['Python_version'] = sys.version
+    options['tool_dependencies']['python_package']['esmecata'] = esmecata_version
+    options['tool_dependencies']['python_package']['pronto'] = pronto.__version__
+    options['tool_dependencies']['python_package']['gseapy'] = gseapy.__version__
+    options['tool_dependencies']['python_package']['orsum'] = get_orsum_version()
+    options['tool_dependencies']['python_package']['pandas'] = pd.__version__
+    options['tool_dependencies']['python_package']['biopython'] = biopython_version
+    options['tool_dependencies']['python_package']['matplotlib'] = matplotlib_version
+
+    esmecata_metadata = {}
+    date = datetime.datetime.now().strftime('%d-%B-%Y %H:%M:%S')
+    esmecata_metadata['access_time'] = date
+    esmecata_metadata['tool_options'] = options
 
     # If no enzyme.Dat available download it.
     if enzyme_data_file is None:
@@ -179,21 +227,22 @@ def taxon_rank_annotation_enrichment(annotation_folder, output_folder, enzyme_da
             df.to_csv(os.path.join(orsum_input_folder, tax_name), sep='\t', index=False)
 
     enriched_elements = {}
-    for result_file in os.listdir('enrichr_module_phylum_output'):
+    for result_file in os.listdir(orsum_input_folder):
         if result_file.endswith('.txt'):
-            df = pd.read_csv(os.path.join('enrichr_module_phylum_output', result_file), sep='\t')
+            df = pd.read_csv(os.path.join(orsum_input_folder, result_file), sep='\t')
             df.set_index('Term', inplace=True)
             base_name = os.path.splitext(os.path.basename(result_file))[0]
             enriched_elements[base_name] = df[df['Adjusted P-value'] < 0.05]['Odds Ratio'].to_dict()
 
     all_elments = set([element for org in enriched_elements for element in enriched_elements[org]])
 
-    with open('enrich_matrix_phylum.tsv', 'w') as open_output_file:
+    enrich_matrix_file = os.path.join(output_folder, 'enrich_matrix.tsv')
+    with open(enrich_matrix_file, 'w') as open_output_file:
         csvwriter = csv.writer(open_output_file, delimiter='\t')
         csvwriter.writerow(['Organism', *all_elments])
         for org in enriched_elements:
             csvwriter.writerow([org, *[enriched_elements[org][element] if element in enriched_elements[org] else 'NA' for element in all_elments]])
-    result_df = pd.read_csv('enrich_matrix_phylum.tsv', sep='\t')
+    result_df = pd.read_csv(enrich_matrix_file, sep='\t')
     result_df.set_index('Organism', inplace=True)
     fig, axes = plt.subplots(figsize=(70,25))
     plt.rc('font', size=4)
@@ -201,3 +250,16 @@ def taxon_rank_annotation_enrichment(annotation_folder, output_folder, enzyme_da
     # Run orsum to filter list of enriched annotations.
     orsum_output_folder = os.path.join(output_folder, 'orsum_output_folder')
     run_orsum(annotation_file_gmt_file, orsum_input_folder, orsum_output_folder)
+
+    endtime = time.time()
+    duration = endtime - starttime
+    esmecata_metadata['esmecata_gseapy_taxon_duration'] = duration
+    gseapy_taxon_metadata_file = os.path.join(output_folder, 'esmecata_metadata_gseapy_taxon.json')
+    if os.path.exists(gseapy_taxon_metadata_file):
+        metadata_files = [metadata_file for metadata_file in os.listdir(output_folder) if 'esmecata_metadata_gseapy_taxon' in metadata_file]
+        gseapy_taxon_metadata_file = os.path.join(output_folder, 'esmecata_metadata_gseapy_taxon_{0}.json'.format(len(metadata_files)))
+        with open(gseapy_taxon_metadata_file, 'w') as ouput_file:
+            json.dump(esmecata_metadata, ouput_file, indent=4)
+    else:
+        with open(gseapy_taxon_metadata_file, 'w') as ouput_file:
+            json.dump(esmecata_metadata, ouput_file, indent=4)